+++ src/pyatmo/modules/device_types.py
@@ -75,6 +75,7 @@ class DeviceType(str, Enum):
     NLUP = "NLUP"  # Legrand device stub
     NLLF = "NLLF"  # Legrand Centralized Ventilation Control
     NLTS = "NLTS"  # Legrand motion sensor stub
+    NLJ = "NLJ"  # Legrand garage door opener
 
     # BTicino Classe 300 EOS
     BNCX = "BNCX"  # internal panel = gateway
@@ -190,6 +191,7 @@ DEVICE_CATEGORY_MAP: dict[DeviceType, DeviceCategory] = {
     DeviceType.BNFC: DeviceCategory.climate,
     DeviceType.BNTR: DeviceCategory.climate,
     DeviceType.NLPD: DeviceCategory.switch,
+    DeviceType.NLJ: DeviceCategory.shutter,
 }
 
 
@@ -246,6 +248,7 @@ DEVICE_DESCRIPTION_MAP: dict[DeviceType, tuple[str, str]] = {
     DeviceType.NLUI: ("Legrand", "In-wall switch"),
     DeviceType.NLTS: ("Legrand", "Motion sensor"),
     DeviceType.NLUF: ("Legrand", "In-Wall dimmer"),
+    DeviceType.NLJ: ("Legrand", "Garage door opener"),
     # BTicino Classe 300 EOS
     DeviceType.BNCX: ("BTicino", "Internal Panel"),
     DeviceType.BNEU: ("BTicino", "External Unit"),